requirements.txt
pandas>=2.0
geopandas>=1.0
shapely>=2.0
pyproj>=3.6
holidays>=0.50
pyarrow>=14.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
scipy>=1.7.0
matplotlib>=3.5.0
PyYAML>=6.0
